+++ private/cli/develop.rkt
@@ -8,19 +8,25 @@
   racket/list
   racket/cmdline
   racket/file
+  setup/getinfo
   unlike-assets/logging
   unlike-assets
   file-watchers
   "../../main.rkt"
   "../../paths.rkt"
+  "../server.rkt"
   "shared.rkt")
 
 (define (develop)
   (define timeout (make-parameter 500))
+  (define port (make-parameter 8080))
 
   (command-line
     #:program "develop"
     #:once-each
+    [("-p" "--port") user-port
+                     "The port to use for the dev server. Default: 8080."
+                     (port (string->number user-port))]
     [("--delay") ms
                  "The number of milliseconds to allow between changes "
                  "before trying to compile again. Default: 500."
@@ -84,11 +90,18 @@
           (build-with-report changed removed))
         (loop)))))
 
+    (define stop
+      (start-server (dist-rel) (port)))
+
+    (define info-version ((get-info/full (polyglot-rel)) 'version))
     (with-handlers ([exn:break? (λ _
                                    (printf "~nStopped.~n")
                                    (kill-thread aggregator)
                                    (kill-thread watcher)
+                                   (stop)
                                    (exit 0))])
-      (displayln "Watching for changes. ^C to stop.")
+      (printf "~n~n~a~nPolyglot ~a~n" (make-string 40 #\=) info-version)
+      (printf "Development server online at http://localhost:~a~n" (port))
+      (printf "Press ^C to stop.~n")
       (thread-wait aggregator)
       (thread-wait watcher))))